project2_attention_cnn/requirements.txt
torch==2.4.0
torchvision==0.19.0
torchaudio==2.4.0
Pillow==10.4.0
numpy==1.26.4
streamlit==1.39.0
tqdm==4.66.5
